test/core-py/requirements.txt
Flask==0.12.2
cfenv==0.5.3
sap_cf_logging==3.0.0
sklearn